Modified - refactor main drawer to core
