+++ lib/core/widgets/main_drawer.dart
@@ -7,15 +7,15 @@ import 'package:tamizshahr/provider/customer_info.dart';
 import 'package:tamizshahr/screens/charity_screen.dart';
 import 'package:tamizshahr/screens/messages_screen.dart';
 
-import '../provider/auth.dart';
-import '../screens/about_us_screen.dart';
-import '../screens/cart_screen.dart';
-import '../screens/contact_with_us_screen.dart';
-import '../features/customer_feature/presentation/screens/login_screen.dart';
-import '../features/customer_feature/presentation/screens/profile_screen.dart';
-import '../screens/guide_screen.dart';
-import '../core/screens/navigation_bottom_screen.dart';
-import '../screens/product_screen.dart';
+import '../../provider/auth.dart';
+import '../../screens/about_us_screen.dart';
+import '../../screens/cart_screen.dart';
+import '../../screens/contact_with_us_screen.dart';
+import '../../features/customer_feature/presentation/screens/login_screen.dart';
+import '../../features/customer_feature/presentation/screens/profile_screen.dart';
+import '../../screens/guide_screen.dart';
+import '../screens/navigation_bottom_screen.dart';
+import '../../screens/product_screen.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
