+++ lib/features/customer_feature/presentation/screens/customer_notification_screen.dart
@@ -3,7 +3,7 @@ import '../../../../screens/messages_screen.dart';
 
 import '../../../../provider/app_theme.dart';
 import '../../../../models/customer.dart';
-import '../../../../widgets/main_drawer.dart';
+import '../../../../core/widgets/main_drawer.dart';
 
 class CustomerNotificationScreen extends StatefulWidget {
   static const routeName = '/customer_notification_screen';
