+++ lib/screens/cart_screen.dart
@@ -15,7 +15,7 @@ import '../widgets/card_item.dart';
 import '../widgets/custom_dialog_enter.dart';
 import '../widgets/custom_dialog_profile.dart';
 import '../widgets/en_to_ar_number_convertor.dart';
-import '../widgets/main_drawer.dart';
+import '../core/widgets/main_drawer.dart';
 
 class CartScreen extends StatefulWidget {
   static const routeName = '/cart_screen';
