+++ lib/screens/charity_screen.dart
@@ -9,7 +9,7 @@ import '../models/search_detail.dart';
 import '../provider/Products.dart';
 import '../provider/app_theme.dart';
 import '../widgets/en_to_ar_number_convertor.dart';
-import '../widgets/main_drawer.dart';
+import '../core/widgets/main_drawer.dart';
 
 class CharityScreen extends StatefulWidget {
   static const routeName = '/charitiesScreen';
