+++ lib/screens/collect_detail_screen.dart
@@ -8,7 +8,7 @@ import 'package:tamizshahr/widgets/collect_details_collects_item.dart';
 import 'package:tamizshahr/widgets/en_to_ar_number_convertor.dart';
 
 import '../provider/app_theme.dart';
-import '../widgets/main_drawer.dart';
+import '../core/widgets/main_drawer.dart';
 
 class CollectDetailScreen extends StatefulWidget {
   static const routeName = '/collectDetailScreen';
