+++ lib/screens/product_detail_screen.dart
@@ -11,7 +11,7 @@ import '../models/product.dart';
 import '../provider/Products.dart';
 import '../provider/app_theme.dart';
 import '../widgets/en_to_ar_number_convertor.dart';
-import '../widgets/main_drawer.dart';
+import '../core/widgets/main_drawer.dart';
 import 'cart_screen.dart';
 
 class ProductDetailScreen extends StatefulWidget {
