+++ lib/screens/waste_cart_screen.dart
@@ -13,7 +13,7 @@ import '../features/waste_feature/presentation/wastes_screen.dart';
 import '../widgets/custom_dialog_enter.dart';
 import '../widgets/custom_dialog_profile.dart';
 import '../widgets/en_to_ar_number_convertor.dart';
-import '../widgets/main_drawer.dart';
+import '../core/widgets/main_drawer.dart';
 import '../widgets/waste_cart_item.dart';
 import 'address_screen.dart';
